% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimaggio_1941}
\alias{dimaggio_1941}
\title{Joe DiMaggio's 1941 season}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139 rows and 38 columns.
}
\usage{
dimaggio_1941
}
\description{
Joe DiMaggio's 1941 season
}
\keyword{datasets}
