% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hof_batting}
\alias{hof_batting}
\alias{hof_pitching}
\title{Hall of Fame batting}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 167 rows and 46 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 70 rows and 30 columns.
}
\usage{
hof_batting

hof_pitching
}
\description{
Hall of Fame batting
}
\keyword{datasets}
