% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_rates.R
\name{hr.rates}
\alias{hr.rates}
\title{Compute home run rates}
\usage{
hr.rates(age, hr, ab)
}
\arguments{
\item{age}{Age of the player}

\item{hr}{Number of hume runs}

\item{ab}{Number of at-bats}
}
\value{
a \code{list} with two elements
}
\description{
Compute home run rates
}
\examples{
hr.rates(35, 4, 200)

}
