% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_rates.R
\name{hr_rates}
\alias{hr_rates}
\title{Compute home run rates}
\usage{
hr_rates(age, hr, ab)
}
\arguments{
\item{age}{Age of the player}

\item{hr}{Number of home runs}

\item{ab}{Number of at-bats}
}
\value{
a \code{list} with two elements
}
\description{
Compute home run rates
}
\examples{
hr_rates(35, 4, 200)

}
