% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.simulation.68.R
\name{one_simulation_68}
\alias{one_simulation_68}
\title{Simulate one season from 1968}
\usage{
one_simulation_68(s.talent = 0.2)
}
\arguments{
\item{s.talent}{Proportion of talent (vs. luck)}
}
\description{
Simulate one season from 1968
}
\examples{
one_simulation_68()

}
