% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retro2016}
\alias{retro2016}
\title{Retrosheet events data for 2016}
\format{
An object of class \code{baseballr_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.table}, \code{data.frame}) with 190713 rows and 98 columns.
}
\usage{
retro2016
}
\description{
Retrosheet events data for 2016
}
\seealso{
\code{\link[baseballr:retrosheet_data]{baseballr::retrosheet_data()}}
}
\keyword{datasets}
