% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retro_gl_header}
\alias{retro_gl_header}
\alias{retro_gl_2011}
\title{Retrosheet game logs}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 0 rows and 161 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2429 rows and 161 columns.
}
\usage{
retro_gl_header

retro_gl_2011
}
\description{
Retrosheet game logs
}
\keyword{datasets}
