% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrosheet.R
\name{retrosheet_add_states}
\alias{retrosheet_add_states}
\alias{retrosheet_add_counts}
\title{Compute Base-Out States on Retrosheet Data}
\source{
\url{https://beanumber.github.io/abdwr3e/05-runsexpectancy.html#sec-runsexpectancy}
}
\usage{
retrosheet_add_states(data, ...)

retrosheet_add_counts(data, ...)
}
\arguments{
\item{data}{A data frame of Retrosheet event data generated by
\code{\link[baseballr:retrosheet_data]{baseballr::retrosheet_data()}}.}

\item{...}{currently ignored}
}
\value{
A tibble with nine additional variables provided information
about the beginning and ending base-out state for the play.
In particular \code{state} is the initial state, and \code{new_state} is
the resulting state.
}
\description{
Compute Base-Out States on Retrosheet Data
}
\examples{
\dontrun{
if (require(baseballr)) {
  retro2016 <- baseballr::retrosheet_data(tempdir(), 2016)
  events <- retro2016[[1]][["events"]]
  events_states <- retrosheet_add_states(events)
}
}
}
\seealso{
\code{\link[baseballr:retrosheet_data]{baseballr::retrosheet_data()}}
}
