% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ripken_gamelog}
\alias{ripken_gamelog}
\title{Ripken's Iron Man game log}
\format{
\describe{
\item{Variable}{Name of the variable}
\item{Value}{Value of the variable}
}
}
\usage{
ripken_gamelog
}
\description{
Ripken's Iron Man game log
}
\seealso{
\link{retro_gl_header}
}
\keyword{datasets}
