% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spahn}
\alias{spahn}
\title{Warren Spahn}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 34 columns.
}
\usage{
spahn
}
\description{
Warren Spahn
}
\keyword{datasets}
