% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_daily}
\alias{statcast_daily}
\alias{statcast_season}
\alias{statcast_read_csv}
\alias{statcast_write_rds}
\title{Download Statcast data from Baseball Savant}
\usage{
statcast_daily(the_date = lubridate::now(), dir = getwd())

statcast_season(year = lubridate::year(lubridate::now()), dir = getwd())

statcast_read_csv(dir = getwd(), pattern = "*.csv")

statcast_write_rds(x, dir = getwd(), ...)
}
\arguments{
\item{the_date}{The date on which you want to download data}

\item{dir}{Path to a directory to store downloaded files}

\item{year}{The year for which you want to download data}

\item{pattern}{passed to \code{\link[base:list.files]{base::list.files()}}}

\item{x}{Data frame to write. See \code{\link[readr:read_rds]{readr::write_rds()}}.}

\item{...}{arguments passed to \code{\link[readr:read_rds]{readr::write_rds()}}. Currently ignored.}
}
\description{
Baseball Savant downloads are limited to 25,000 rows. These functions help
you build a larger database easily.
}
\examples{
statcast_daily("2023-06-11", dir = tempdir())

x <- statcast_read_csv(tempdir())
x <- statcast_read_csv(tempdir())
statcast_write_rds(x, tempdir())

}
\seealso{
\code{\link[baseballr:statcast_search]{baseballr::statcast_search()}}
}
