% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temps_2023}
\alias{temps_2023}
\alias{parks_2023}
\title{Ballparks and temperatures}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2199 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 2 columns.
}
\usage{
temps_2023

parks_2023
}
\description{
Ballparks and temperatures
}
\keyword{datasets}
