% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabrera}
\alias{cabrera}
\alias{sanchez}
\alias{umpires}
\alias{verlander}
\title{Ball-Strike Count data}
\format{
An object of class \code{data.frame} with 6265 rows and 12 columns.

An object of class \code{data.frame} with 9113 rows and 10 columns.

An object of class \code{data.frame} with 378365 rows and 9 columns.

An object of class \code{data.frame} with 15307 rows and 12 columns.
}
\usage{
cabrera

sanchez

umpires

verlander
}
\description{
Ball-Strike Count data
}
\details{
Pitch-by-pitch level detail for several players and umpires.
}
\seealso{
\url{https://beanumber.github.io/abdwr3e/06-pitchcount.html#behaviors-by-count}
}
\keyword{datasets}
