% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{compareTrajectories}
\alias{compareTrajectories}
\title{Shiny app for comparing player trajectories}
\usage{
compareTrajectories()
}
\description{
Shiny app for comparing player trajectories
}
\examples{
\dontrun{
  compareTrajectories()
}
}
