% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimaggio_1941}
\alias{dimaggio_1941}
\alias{williams_1941}
\title{Ted Williams' and Joe DiMaggio's 1941 seasons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139 rows and 38 columns.

\describe{
\item{Rk}{A ranking index}
\item{Gcar}{Cumulative number of career games}
\item{Gtm}{game number}
\item{Date}{date of game}
\item{Tm}{player team}
\item{X6}{indicates home or awy}
\item{Opp}{opposing team}
\item{Rslt}{game outcome}
\item{Inngs}{number of innings played}
\item{PA}{plate appearances}
\item{AB}{at-bats}
\item{R}{runs scored}
\item{H}{hits}
\item{X2B}{doubles}
\item{X3B}{triples}
\item{HR}{home runs}
\item{RBI}{runs batted in}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitches}
\item{SH}{sacrifice hits}
\item{SF}{sacrifice flies}
\item{ROE}{reached on an error}
\item{GDP}{grounded in double play}
\item{SB}{stolen bases}
\item{CS}{caught stealing}
\item{BA}{batting average}
\item{OBP}{on-base percentage}
\item{SLG}{slugging percentage}
\item{OPS}{ops statistic}
\item{BOP}{base-out percentage}
\item{Pos}{fielding position}
}
}
\source{
Baseball-Reference: \url{https://www.baseball-reference.com/players/gl.fcgi?id=dimagjo01&t=b&year=1941}

Baseball-Reference: \url{https://www.baseball-reference.com/players/gl.fcgi?id=willite01&t=b&year=1941}
}
\usage{
dimaggio_1941

williams_1941
}
\description{
Game-by-game hitting data for Joe DiMaggio in the 1941 season

Game-by-game hitting data for Ted Williams in the 1941 season
}
\keyword{datasets}
