% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hof_batting}
\alias{hof_batting}
\alias{hof_pitching}
\title{Hall of Fame Batting and Pitching Statistics}
\format{
\describe{
\item{Player}{player name}
\item{HR}{career home runs}
\item{AB}{career at-bats}
\item{From}{first season in MLB}
\item{To}{last season in MLB}
\item{OBP}{career on-base percentage}
\item{SLG}{career slugging percentage}
\item{OPS}{career OPS statistic}
}

\describe{
\item{Rk}{A ranking number}
\item{X2}{player name}
\item{Inducted}{year inducated in HOF}
\item{Yrs}{length of career}
\item{From}{first year in MLB}
\item{To}{final year in MLB}
\item{ASG}{number of all-star game appearances}
\item{WAR}{career WAR measure}
\item{W}{career wins}
\item{L}{career losses}
\item{W.L.}{winning fraction}
\item{ERA}{career ERA}
\item{G}{career games}
\item{GS}{career games started}
\item{GF}{career games finished}
\item{CG}{complete games}
\item{SHO}{shutouts}
\item{SV}{saves}
\item{IP}{innings pitched}
\item{H}{hits allowed}
\item{R}{runs allowed}
\item{ER}{earned runs allowed}
\item{HR}{home runs allowed}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitches}
\item{BK}{balks}
\item{WP}{wild pitches}
\item{BF}{batters faced}
}
}
\source{
Baseball-Reference
}
\usage{
hof_batting

hof_pitching
}
\description{
Career batting and pitching statistics for players inducted in the
baseball Hall of Fame
}
\seealso{
\link[Lahman:HallOfFame]{Lahman::HallOfFame}
}
\keyword{datasets}
