% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLbatting}
\alias{NLbatting}
\alias{ALbatting}
\alias{NLpitching}
\title{American and National League statistics}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 28 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 28 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 34 columns.
}
\usage{
NLbatting

ALbatting

NLpitching
}
\description{
American and National League statistics
}
\keyword{datasets}
