% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_simulation_68.R
\name{one_simulation_68}
\alias{one_simulation_68}
\title{Simulate one season from 1968}
\usage{
one_simulation_68(s.talent = 0.2)
}
\arguments{
\item{s.talent}{Proportion of talent (vs. luck)}
}
\value{
A \code{tbl} with 20 rows (one for each team). Each call to function
simulates one random season from 1968.
}
\description{
Simulate one season from 1968
}
\examples{
one_simulation_68()

}
