% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ripken_gamelog}
\alias{ripken_gamelog}
\title{Ripken's Iron Man game}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 48 rows and 2 columns.
}
\usage{
ripken_gamelog
}
\description{
Ripken's Iron Man game
}
\keyword{datasets}
