% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spahn}
\alias{spahn}
\title{Warren Spahn Pitching Statistics}
\format{
\describe{
\item{Year}{season}
\item{Age}{age}
\item{Tm}{team}
\item{Lg}{league}
\item{W}{number of wins}
\item{L}{number of losses}
\item{W.L}{winning fraction}
\item{ERA}{earned run average}
\item{G}{games}
\item{GS}{games started}
\item{GF}{games finished}
\item{CG}{complete games}
\item{SHO}{shutouts}
\item{SV}{saves}
\item{IP}{innings pitched}
\item{H}{hits}
\item{R}{runs}
\item{ER}{earned runs}
\item{HR}{home runs allowed}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitch}
\item{BK}{balks}
\item{WP}{wild pitches}
\item{BF}{batters faced}
\item{ERA.}{earned run average}
\item{WHIP}{walks and hits per nine innings}
\item{H.9}{hits per nine innings}
\item{HR.9}{home runs per nine innings}
\item{BB.9}{walks per nine innings}
\item{SO.9}{strikeouts per nine innings}
\item{SO.BB}{ratio of strikeouts to walks}
}
}
\usage{
spahn
}
\description{
Season pitching statistics for the Hall of Fame pitcher Warren Spahn
}
\keyword{datasets}
