% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temps_2023}
\alias{temps_2023}
\alias{parks_2023}
\title{Ballpark names and game time temperatures for 2023}
\format{
\describe{
\item{game_pk}{MLBAM game identifier}
\item{temperature}{Game time temperature in degrees Fahrenheit}
}

\describe{
\item{Park}{The name of the ballpark}
\item{Dome}{Does the park have a domed roof?}
}
}
\usage{
temps_2023

parks_2023
}
\description{
Ballpark names and game time temperatures for 2023
}
\keyword{datasets}
